test_events.sh: test equations from 'events_list.txt' file
This is the testing file. It reads lines from 'events_list.txt', uses
'parse_args.pl' to run 'perf stat' command and then creates an equation
by using 'parse_stat.pl'. This equation is then tested and the result
is printed to stdout alongside with the line from 'events_list.txt' and
with the equation from 'parse_stat.pl'.